+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## master
 
 * The `BottomBannerViewController` now accounts for the safe-area inset correctly, if applicable. ([#1982](https://github.com/mapbox/mapbox-navigation-ios/pull/1982))
+
+### CarPlay
+
+* `CarPlayNavigationViewController` now exposes the `navigationService`. ([#2005](https://github.com/mapbox/mapbox-navigation-ios/pull/2005))
+* `CarPlayNavigationViewControllerDelegate` now has a message for when a user arrives at a waypoint. ([#2005](https://github.com/mapbox/mapbox-navigation-ios/pull/2005))
 * `CarPlayManager` will now notify its delegate if the route request fails and provide the option to present an alert on the map template. ([#1981](https://github.com/mapbox/mapbox-navigation-ios/pull/1981))
 
 ## v0.29.1